+++ config.yml
@@ -1,22 +1,15 @@
-language: hu
+language: en
 
 pipeline:
   - name: SpacyNLP
   - name: SpacyTokenizer
   - name: SpacyFeaturizer
   - name: CountVectorsFeaturizer
-  - name: services.hun_date_extractor.HunDateExtractor
   - name: DIETClassifier
     epochs: 50
     entity_recognition: False
-  - name: FallbackClassifier
-    threshold: 0.35
   - name: ResponseSelector
     epochs: 50
 
 policies:
-   - name: MemoizationPolicy
-   - name: TEDPolicy
-     max_history: 5
-     epochs: 100
    - name: RulePolicy
