+++ data/nlu.yml
@@ -1,4 +1,5 @@
 version: "2.0"
+
 nlu:
 - intent: udvozol
   examples: |
@@ -21,94 +22,3 @@ nlu:
     - esetleg pénteken találkozni?
     - konzultálni szeretnék
     - időpontot foglalnék
-- intent: recommend_date_or_time
-  examples: |
-    - bármikor
-    - ráéreg egész nap
-    - szabad vagyok, amikor neked jó
-    - amikor te ráérsz
-    - amikor csak neked jó
-    - rugalmas vagyok
-    - neked mikor lenne jó?
-    - nekem bármikor jó.
-    - milyen opciók vannak?
-    - tudnál ajánlani egy időpontot?
-    - mikor tudnánk konzultálni?
-    - esetleg tudnánk valamikor találkozni?
-    - mikor érnél rá?
-    - ráérsz valamikor?
-    - mikor érnél rá egy találkozóra?
-    - mikor érnél rá egy konzultálni?
-    - szabad vagy mostanában valamikor?
-    - mikor vagy szabad?
-- intent: another_appointment
-  examples: |
-    - legyen másik nap akkor
-    - akkor egy másik nap esetleg?
-    - akkor legyen inkább a szerda
-    - tegyük egy másik napra
-    - legyen inkább a szerda
-    - tegyük inkább máskorra légyszi
-    - legyen inkább máskor kérlek
-    - mégse jó a hétfő
-    - nem érek rá akkor
-- intent: say_date_or_time
-  examples: |
-    - kedden
-    - vasárnap
-    - holnap
-    - szerdán
-    - december 5-én
-    - jövő kedden
-    - két hét múlva szerdán
-    - november 16-án jó lenne
-    - holnap megfelelne
-    - nekem jó a holnap
-    - kedden ráérsz?
-    - a szerda jó esetleg?
-    - egy találkozó csütörtökön?
-    - holnap ráérek
-    - most
-    - most ráérek
-    - most jó
-    - akár most is jó nekem
-    - nyolc órakor
-    - reggel
-    - délelőtt
-    - 6 órakor
-    - tíz órakor
-    - egész reggel jó
-    - legyen két órakor akkor
-    - ha jó neked is akkor legyen a két óra
-    - akkor legyen 3 órakor
-    - három órakor jó?
-    - két órakor ráérsz?
-    - reggel nyolckor megfelel?
-    - reggel ráérsz?
-    - délután jó?
-    - legyen 16:20-kor
-    - legyen délután négykor
-    - este 7
-    - legyen 12:00
-    - 20:00
-    - 14 óra 46 perckor
-    - délután fél 2 után 5 perccel
-    - 15:23-kor ott vagyok
-    - 17 óra 34-kor ott leszek
-- intent: vege
-  examples: |
-    - oké
-    - oké, köszi
-    - köszi
-    - rendben
-    - ott leszek
-    - akkor ezt megbeszéltük
-    - viszlát
-    - rendben. szia!
-- intent: out_of_scope
-  examples: |
-    - elnök szeretnék lenni
-    - programozni akarok
-    - meg szeretnék tanulni angolul
-    - egy bigmacet szeretnék
-    - megennék egy magyaros pizzát
