+++ data/rules.yml
@@ -1,46 +1,12 @@
 version: "2.0"
 
 rules:
+  - rule: Hi rule
+    steps:
+        - intent: udvozol
+        - action: utter_udvozol
 
-- rule: fallback
-  steps:
-    - intent: nlu_fallback
-    - action: utter_rephrase
-
-- rule: out_of_scope
-  steps:
-    - intent: out_of_scope
-    - action: utter_out_of_scope
-
-- rule: udvozol
-  steps:
-  - intent: udvozol
-  - action: utter_udvozol
-
-- rule: Activate form 1
-  steps:
-  - intent: idopontfoglalas
-  - action: validate_idopont_form
-
-- rule: Activate form 2
-  steps:
-  - intent: say_date_or_time
-  - action: validate_idopont_form
-
-
-- rule: Activate form 3
-  steps:
-  - intent: recommend_date_or_time
-  - action: action_recommend_date
-
-- rule: another appointment
-  steps:
-    - intent: another_appointment
-    - action: action_remove_appointment
-    - action: validate_idopont_form
-
-- rule: viszlat
-  steps:
-    - intent: vege
-    - action: action_remove_appointment
-    - action: utter_bye
+  - rule: Single rule
+    steps:
+      - intent: idopontfoglalas
+      - action: action_fill_table
