+++ requirements.txt
@@ -1,6 +1,8 @@
-rasa
+rasa==2.2.7
 spacy==2.2.1
 fuzzywuzzy
-hun-date-parser==0.0.6
-git+https://github.com/oroszgy/spacy-hungarian-models/releases/download/hu_core_ud_lg-0.3.1/hu_core_ud_lg-0.3.1-py3-none-any.whl
+hun-date-parser==0.0.7
+# git+https://github.com/oroszgy/spacy-hungarian-models/releases/download/hu_core_ud_lg-0.3.1/hu_core_ud_lg-0.3.1-py3-none-any.whl
 datetimerange
+pandas
+altair
